Move ApplicationCommandUtilities into it's own file
